:shirt: Extract Changes into its own file
