+++ lib/owner-notifier.js
@@ -1,4 +1,5 @@
 const ownersFile = require('owners-file');
+const Changes = require('./changes');
 
 class OwnerNotifier {
   constructor(github, event) {
@@ -45,19 +46,6 @@ class OwnerNotifier {
         body: '/cc ' + changes.owners.join(' ')
       }, this.repo));
     }
-  }
-}
-
-class Changes {
-  constructor(paths, ownersFile) {
-    this.paths = paths;
-    this.ownersFile = ownersFile;
-  }
-
-  get owners() {
-    return Array.from(new Set(this.paths.reduce((result, path) => {
-      return result.concat(this.ownersFile.for(path));
-    }, [])));
   }
 }
